timeShift/requirements.txt
numpy==1.26.0
paho_mqtt==1.6.1
pandas==2.0.0
Requests==2.31.0
